@@ -12,24 +12,25 @@
         "correct": "npm run format && npm run lint"
     },
     "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.36",
-        "@fortawesome/free-regular-svg-icons": "^5.15.4",
-        "@fortawesome/free-solid-svg-icons": "^5.15.4",
-        "@fortawesome/react-fontawesome": "^0.1.16",
-        "@material-ui/core": "^4.12.3",
-        "@material-ui/icons": "^4.11.2",
-        "@material-ui/styles": "^4.11.4",
-        "@reduxjs/toolkit": "^1.6.1",
+        "@emotion/react": "^11.10.6",
+        "@emotion/styled": "^11.10.6",
+        "@fortawesome/fontawesome-svg-core": "^6.4.0",
+        "@fortawesome/free-regular-svg-icons": "^6.4.0",
+        "@fortawesome/free-solid-svg-icons": "^6.4.0",
+        "@fortawesome/react-fontawesome": "^0.2.0",
+        "@mui/icons-material": "^5.11.16",
+        "@mui/material": "^5.12.0",
+        "@reduxjs/toolkit": "^1.9.3",
         "muuri": "^0.9.5",
-        "muuri-react": "^3.1.6",
-        "react": "^16.13.0",
-        "react-bootstrap": "^2.0.3",
-        "react-dom": "^16.13.0",
-        "react-redux": "^7.2.4",
-        "react-router-dom": "^6.0.2",
-        "react-scripts": "4.0.3",
+        "react": "^18.2.0",
+        "react-bootstrap": "^2.7.4",
+        "react-dom": "^18.2.0",
+        "react-redux": "^8.0.5",
+        "react-router-dom": "^6.10.0",
+        "react-scripts": "^5.0.1",
         "redux-persist": "^6.0.0",
-        "swr": "^1.1.2"
+        "ruuri": "^1.0.0",
+        "swr": "^2.1.3"
     },
     "devDependencies": {
         "@types/jest": "^27.4.1",
@@ -38,7 +39,10 @@
         "@types/spotify-api": "^0.0.21",
         "prettier": "2.8.7",
         "sass": "^1.62.0",
-        "typescript": "^4.5.5"
+        "typescript": "^5.0.4"
+    },
+    "overrides": {
+        "typescript": "^5.0.4"
     },
     "eslintConfig": {
         "extends": "react-app"
